improved the architechure of code generation and the accuracy of it
